+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## **master** (MELPA)
 - Added support for `Verb-Map-Response` heading property. This allows for calling functions automatically with the HTTP response data, when it is received.
 - Added the `verb-util-form-url-encode` helper function for use with `application/x-www-form-urlencoded`.
+- Fixed error occuring when maxium number of redirections set by `url-max-redirections` was reached.
+- The `verb-headers-get` function now accepts an instance of `verb-response` as first argument.
 
 ## **3.0.0** - 2024-12-23 (MELPA Stable)
 Breaking changes: